pyproject.toml
[tool.poetry]
name = "temp-api"
version = "0.1.0"
description = ""
authors = ["mutwiriian <iiyanmutuma@gmail.com>"]
readme = "README.md"
packages = [{include = "temp_api"}]

[tool.poetry.dependencies]
python = "^3.10"
ipykernel = "^6.22.0"
scikit-learn = "^1.2.2"
numpy = "^1.24.2"
pandas = "^2.0.0"
lightgbm = "^3.3.5"
matplotlib = "^3.7.1"
joblib = "^1.2.0"
fastapi = "^0.95.1"
ipython = "^8.12.0"
pydantic = {extras = ["email"], version = "^1.10.7"}
email-validator = "^2.0.0.post1"
uvicorn = "^0.21.1"
sqlalchemy = "^2.0.13"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptograpghy"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
